requirements.txt
raven==6.8.0
mysqlclient==1.3.12
retrying==1.3.3
PyYAML==3.12
